+++ src/cpu/instructions/logical.rs
@@ -1,56 +1,56 @@
 use crate::cpu::Cpu;
 
 impl Cpu {
-    
-/// AND Accumulator with Memory (DP Indexed Indirect,X)
-pub fn exe_and(&mut self, data: u32) {
-	todo!()
-}
-
-/// Test Bits (Direct Page)
-pub fn exe_bit(&mut self, data: u32) {
-	todo!()
-}
-
-/// Compare Accumulator with Memory (DP Indexed Indirect,X)
-pub fn exe_cmp(&mut self, data: u32) {
-	todo!()
-}
-
-/// Compare Index Register X with Memory (Immediate)
-pub fn exe_cpx(&mut self, data: u32) {
-	todo!()
-}
-
-/// Compare Index Register Y with Memory (Immediate)
-pub fn exe_cpy(&mut self, data: u32) {
-	todo!()
-}
-
-/// Exclusive-OR Accumulator with Memory (DP Indexed Indirect,X)
-pub fn exe_eor(&mut self, data: u32) {
-	todo!()
-}
-
-/// Logical Shift Memory or Accumulator Right (Direct Page)
-pub fn exe_lsr(&mut self, data: u32) {
-	todo!()
-}
-
-/// OR Accumulator with Memory (DP Indexed Indirect,X)
-pub fn exe_ora(&mut self, data: u32) {
-	todo!()
-}
-
-/// Rotate Memory or Accumulator Left (Direct Page)
-pub fn exe_rol(&mut self, data: u32) {
-	todo!()
-}
-
-/// Rotate Memory or Accumulator Right (Direct Page)
-pub fn exe_ror(&mut self, data: u32) {
-	todo!()
-}
-
-
+	
+	/// AND Accumulator with Memory (DP Indexed Indirect,X)
+	pub fn exe_and(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Test Bits (Direct Page)
+	pub fn exe_bit(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Compare Accumulator with Memory (DP Indexed Indirect,X)
+	pub fn exe_cmp(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Compare Index Register X with Memory (Immediate)
+	pub fn exe_cpx(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Compare Index Register Y with Memory (Immediate)
+	pub fn exe_cpy(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Exclusive-OR Accumulator with Memory (DP Indexed Indirect,X)
+	pub fn exe_eor(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Logical Shift Memory or Accumulator Right (Direct Page)
+	pub fn exe_lsr(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// OR Accumulator with Memory (DP Indexed Indirect,X)
+	pub fn exe_ora(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Rotate Memory or Accumulator Left (Direct Page)
+	pub fn exe_rol(&mut self, data: u16) {
+		todo!()
+	}
+	
+	/// Rotate Memory or Accumulator Right (Direct Page)
+	pub fn exe_ror(&mut self, data: u16) {
+		todo!()
+	}
+	
+	
 }
